% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ssurgo_portal}
\alias{install_ssurgo_portal}
\title{Download SSURGO Portal .PYZ file from GitHub}
\usage{
install_ssurgo_portal(verbose = TRUE, overwrite = FALSE, timeout = 3000, ...)
}
\arguments{
\item{verbose}{Show download progress and location of file on successful download? Default `TRUE`}

\item{overwrite}{Overwrite existing PYZ file? Default: `FALSE`}

\item{timeout}{Default: `3000` seconds. Temporarily overrides `options()` for `timeout`.}

\item{...}{Additional arguments to `download.file()`}
}
\value{
Path to downloaded file, or `try-error` on error.
}
\description{
This routine downloads the latest .pyz file from: \url{https://github.com/ncss-tech/SSURGO-Portal/}
}
\examples{
\dontrun{
 install_ssurgo_portal()
}
}
