% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\name{SSURGOPORTAL_PYTHON_VERSION}
\alias{SSURGOPORTAL_PYTHON_VERSION}
\alias{SSURGOPORTAL_GDAL_VERSION}
\title{SSURGOPortal Environment Variables}
\usage{
SSURGOPORTAL_PYTHON_VERSION()

SSURGOPORTAL_GDAL_VERSION()
}
\value{
character. Version number for Python or GDAL to be used when building Python environment.
}
\description{
\code{SSURGOPORTAL_PYTHON_VERSION()}: returns value of system environment
variable \code{R_SSURGOPORTAL_PYTHON_VERSION} or option \code{SSURGOPortal.python_version}.
If neither are set then returns \code{"3.10.2"}, the suggested Python version.
}
