% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ssurgo_portal_dir}
\alias{ssurgo_portal_dir}
\title{SSURGO Portal User Directories}
\usage{
ssurgo_portal_dir(which = "data")
}
\arguments{
\item{which}{One or more of: \code{"data"}, \code{"config"}, \code{"cache"}}
}
\value{
\code{character}. Path to data, config or cache directories, respectively.
}
\description{
These are standard locations for storing data, config, and cache information for SSURGO Portal outside of the installation in the R package library.
}
\examples{
ssurgo_portal_dir("data")
ssurgo_portal_dir("config")
ssurgo_portal_dir("cache")
}
