% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venv.R
\name{create_ssurgo_venv}
\alias{create_ssurgo_venv}
\alias{ssurgo_portal_python}
\title{Create Virtual Environment with SSURGO Portal Dependencies}
\usage{
create_ssurgo_venv(
  envname = "r-ssurgoportal",
  version = "3.10:latest",
  gdal_version = "3.7.3"
)

ssurgo_portal_python(
  envname = getOption("SSURGOPortalR.virtualenv_name", default = "r-ssurgoportal")
)
}
\arguments{
\item{envname}{\emph{character}. Python virtual environment name. Default: \code{"r-ssurgoportal"}}

\item{version}{\emph{character}. Semantic version number of 'Python' to install if not available. Default: \code{"3.10:latest"}, the latest version of Python 3.10.x.}

\item{gdal_version}{\emph{character}. Semantic version number of 'GDAL' package to install. Default: \code{"3.7.3"}}
}
\value{
\emph{character}. Path to virtual environment Python binary (invisible).
}
\description{
Uses reticulate to create a virtual environment with the SSURGO Portal dependencies. Select the environment name, default \code{"r-ssurgoportal"}, along with the specific Python and GDAL versions.
}
