% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venv.R
\name{create_ssurgo_venv}
\alias{create_ssurgo_venv}
\alias{ssurgo_portal_python}
\title{Create Virtual Environment with SSURGO Portal Dependencies}
\usage{
create_ssurgo_venv(
  envname = "r-ssurgoportal",
  python_version = SSURGOPORTAL_PYTHON_VERSION(),
  gdal_version = SSURGOPORTAL_GDAL_VERSION(),
  ...
)

ssurgo_portal_python(
  envname = getOption("SSURGOPortal.virtualenv_name", default = "r-ssurgoportal"),
  ...
)
}
\arguments{
\item{envname}{\emph{character}. Python virtual environment name. Default: \code{"r-ssurgoportal"}}

\item{python_version}{\emph{character}. Semantic version number of 'Python' to install if not available. Default: `"3.10.2"}

\item{gdal_version}{\emph{character}. Semantic version number of 'GDAL' package to install. Default: \code{""} for no specific version.}

\item{...}{Additional arguments not (currently) used.}
}
\value{
\emph{character}. Path to virtual environment Python binary (invisible).
}
\description{
Uses 'reticulate' to create a virtual environment with the SSURGO Portal
dependencies. Select the environment name, default \code{"r-ssurgoportal"},
along with the specific Python and GDAL versions.
}
