% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ssurgo_portal}
\alias{install_ssurgo_portal}
\title{Download SSURGO Portal .PYZ file from GitHub}
\usage{
install_ssurgo_portal(
  verbose = TRUE,
  overwrite = FALSE,
  timeout = 3000,
  src = "https://github.com/ncss-tech/SSURGO-Portal/raw/main/SSURGO-Portal.pyz",
  ...
)
}
\arguments{
\item{verbose}{Show download progress and location of file on successful download? Default \code{TRUE}}

\item{overwrite}{Overwrite existing .PYZ file? Default: \code{FALSE}}

\item{timeout}{Default: \code{3000} seconds. Temporarily overrides \code{options()} for \code{timeout}.}

\item{src}{Default: \code{"https://github.com/ncss-tech/SSURGO-Portal/raw/main/SSURGO-Portal.pyz"}}

\item{...}{Additional arguments to \code{download.file()}}
}
\value{
Path to downloaded file, or \code{try-error} on error.
}
\description{
This routine downloads the latest .pyz file from: \url{https://github.com/ncss-tech/SSURGO-Portal/}
}
\examples{
\dontrun{
 install_ssurgo_portal()
}
}
