% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ssurgo_portal}
\alias{install_ssurgo_portal}
\title{Download SSURGO Portal .PYZ file from Web Soil Survey}
\usage{
install_ssurgo_portal(
  verbose = TRUE,
  overwrite = FALSE,
  timeout = 3000,
  src =
    "https://websoilsurvey.sc.egov.usda.gov/DSD/Download/SsurgoPortal/SSURGO_Portal.zip",
  envname = "r-ssurgoportal",
  python_version = SSURGOPORTAL_R_PYTHON_VERSION(),
  gdal_version = SSURGOPORTAL_R_GDAL_VERSION(),
  ...
)
}
\arguments{
\item{verbose}{Show download progress and location of file on successful download? Default \code{TRUE}}

\item{overwrite}{Overwrite existing .PYZ file? Default: \code{FALSE}}

\item{timeout}{Default: \code{3000} seconds. Temporarily overrides \code{options()} for \code{timeout}.}

\item{src}{Default: \code{"https://websoilsurvey.sc.egov.usda.gov/DSD/Download/SsurgoPortal/SSURGO_Portal.zip"}}

\item{envname}{Virtual environment to create for installation. Default: \code{"r-ssurgoportal"}. Use \code{""} for no virtual environment modifications.}

\item{python_version}{\emph{character}. Semantic version number of 'Python' to install if not available. Default: \code{"3.10.2"}}

\item{gdal_version}{\emph{character}. Semantic version number of 'GDAL' package to install. Default: \code{""}}

\item{...}{Additional arguments to \code{create_ssurgo_venv()} when \code{venv=TRUE}.}
}
\value{
Path to downloaded file, or \code{try-error} on error.
}
\description{
This routine downloads the latest .pyz file from Web Soil Survey
}
\examples{
\dontrun{
 install_ssurgo_portal()
}
}
