% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssurgo_portal.R
\name{ssurgo_portal}
\alias{ssurgo_portal}
\title{Generate Requests for 'SSURGO Portal' Data Loader}
\usage{
ssurgo_portal(
  request = NULL,
  pyz_path = file.path(ssurgo_portal_dir("data"), "SSURGOPortal.pyz"),
  ...,
  schema = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{request}{One of the available request types, see Details.}

\item{pyz_path}{Path to "SSURGO Portal" .pyz file. Default \code{"SSURGOPortal.pyz"} in \code{ssurgo_portal_dir("data")} directory.}

\item{...}{Additional parameters for \code{request}}

\item{schema}{Return request and response schemas? Default: \code{FALSE}}

\item{command_only}{Return command string rather than executing it? Default: \code{FALSE}}
}
\value{
A \code{list} object corresponding to the JSON response. \code{NULL} (invisibly) on error along with any other standard output from the tool printed to the console.
}
\description{
Generate commands for the 'SSURGO Portal' Data Loader command-line interface, and parse resulting JSON output.
}
\details{
\tabular{ll}{
   Request Name \tab Description \cr
   getstatus \tab Quick check for application responsiveness \cr
   getusage \tab Return usage information in payload. \cr
   getwindowsdriveletters \tab Return all drive letters (only valid for Microsoft Windows) \cr
   getfoldertree \tab Returns a file system tree. \cr
   gettemplatecatalog \tab Returns JSON containing information about all available empty SSURGO SQLite templates. \cr
   copytemplatefile \tab Copies a template file to a specified folder path and name. \cr
   opentemplate \tab Opens a SQLite file to confirm that it meets certain minimal criteria. \cr
   getdatabaseinventory \tab List survey areas and related data within a SQLite database. \cr
   deleteareasymbols \tab Delete the specified areasymbols from the database. \cr
   pretestimportcandidates \tab Perform a 'pre-test' on subfolders under a root folder. \cr
   importcandidates \tab Import SSURGO data from subfolders under a root folder. The import terminates if any folder fails. \cr
   importspatialdata \tab For internal use only. Import SSURGO spatial data from shapefiles under a specified path. Note that this activity is isolated to support its use in a child process. \cr
}
}
\examples{
\dontrun{
ssurgo_portal("getstatus")

ssurgo_portal("getstatus", schema = TRUE)

ssurgo_portal(
  "getfoldertree",
  path = "C:/SSURGO Portal/WV",
  folderpattern = "WV",
  filepattern = "txt",
  ignorefoldercase = TRUE,
  ignorefilecase = TRUE,
  showfiles = FALSE,
  maxdepth = 2
)

ssurgo_portal("gettemplatecatalog")

ssurgo_portal(
  "copytemplate",
  templatename = "GeoPackage",
  folder = "test",
  filename = "test.gpkg",
  overwrite = TRUE
)

dbp <- "C:/SSURGO Portal/Databases/West_Virginia_Geopackage.gpkg"

ssurgo_portal("opentemplate", database = dbp)

ssurgo_portal(
  "pretestimportcandidates",
  database = dbp,
  root = "C:/SSURGO Portal/WV",
  istabularonly = FALSE
)

ssurgo_portal(
  "importcandidates",
  database = dbp,
  root = "C:/SSURGO Portal/WV",
  istabularonly = FALSE,
  skippretest = TRUE,
  subfolders = list("WV603"),
  loadinspatialorder = FALSE,
  loadspatialdatawithinsubprocess = TRUE,
  dissolvemupolygon = TRUE
)

ssurgo_portal("getdatabaseinventory", database = dbp)

ssurgo_portal("deleteareasymbols",
              database = dbp,
              areasymbols = list("WV603"))
}
}
